fix: moving sharepoint dependency files to new path
