+++ libs/tools/sharepoint/login.js
