+++ libs/tools/sharepoint/msal.js
