+++ libs/tools/sharepoint/state.js
