+++ tools/sidekick/config.json
@@ -113,6 +113,17 @@
       "environments": ["dev","preview", "live", "prod"],
       "event": "send-to-caas",
       "excludePaths": ["/tools/caas**", "/tools/ost**", "*.json"]
+    },
+    {
+      "containerId": "tools",
+      "id": "version-history",
+      "title": "Version History",
+      "environments": [ "edit" ],
+      "url": "tools/version-history?project=milo--adobecom",
+      "isPalette": true,
+      "passReferrer": true,
+      "paletteRect": "top: auto; bottom: 20px; left: 20px; height: 498px; width: 460px;",
+      "includePaths": [ "**.docx**" ]
     }
   ]
 }
